+++ src/backend/ci/core/common/common-service/build.gradle.kts
@@ -27,7 +27,6 @@
 
 dependencies {
     api(project(":core:common:common-api"))
-    api(project(":core:common:common-security"))
     api(project(":core:common:common-redis"))
     api("org.springframework.boot:spring-boot-starter-actuator")
     api("org.springframework.boot:spring-boot-starter-logging")
