+++ src/backend/ci/core/common/common-service/src/main/kotlin/com/tencent/devops/common/service/utils/CommonUtils.kt
@@ -27,27 +27,19 @@
 
 package com.tencent.devops.common.service.utils
 
-import com.fasterxml.jackson.core.type.TypeReference
-import com.tencent.devops.common.api.auth.AUTH_HEADER_DEVOPS_JWT_TOKEN
 import com.tencent.devops.common.api.constant.CommonMessageCode
 import com.tencent.devops.common.api.exception.ErrorCodeException
-import com.tencent.devops.common.api.pojo.Result
 import com.tencent.devops.common.api.util.JsonUtil
-import com.tencent.devops.common.api.util.MessageUtil
-import com.tencent.devops.common.api.util.OkhttpUtils
-import com.tencent.devops.common.security.jwt.JwtManager
 import com.tencent.devops.common.service.PROFILE_DEFAULT
 import com.tencent.devops.common.service.PROFILE_DEVELOPMENT
 import com.tencent.devops.common.service.PROFILE_PRODUCTION
 import com.tencent.devops.common.service.PROFILE_TEST
 import com.tencent.devops.common.service.Profile
-import java.io.File
 import java.net.Inet4Address
 import java.net.InetAddress
 import java.net.NetworkInterface
 import java.net.SocketException
 import java.util.Enumeration
-import okhttp3.Response
 import org.apache.commons.lang3.StringUtils
 import org.jooq.DSLContext
 import org.slf4j.LoggerFactory
@@ -139,92 +131,6 @@ object CommonUtils {
                 }
             }
         }
-    }
-
-    fun uploadFileToArtifactories(
-        userId: String,
-        serviceUrlPrefix: String,
-        file: File,
-        fileChannelType: String,
-        staticFlag: Boolean = false,
-        language: String,
-        fileRepoPath: String? = null
-    ): Result<String?> {
-        var serviceUrl = "$serviceUrlPrefix/service/artifactories/file/upload" +
-            "?userId=$userId&fileChannelType=$fileChannelType&staticFlag=$staticFlag"
-        fileRepoPath?.let {
-            serviceUrl += "&filePath=$fileRepoPath"
-        }
-        logger.info("serviceUploadFile serviceUrl is:$serviceUrl")
-        uploadFileToService(serviceUrl, file).use { response ->
-            val responseContent = response.body!!.string()
-            logger.error("uploadFile responseContent is: $responseContent")
-            if (!response.isSuccessful) {
-                val message = MessageUtil.getMessageByLocale(
-                    messageCode = CommonMessageCode.SYSTEM_ERROR,
-                    language = language
-                )
-                Result(CommonMessageCode.SYSTEM_ERROR.toInt(), message, null)
-            }
-            return JsonUtil.to(responseContent, object : TypeReference<Result<String?>>() {})
-        }
-    }
-
-    fun uploadFileToService(
-        url: String,
-        uploadFile: File,
-        headers: MutableMap<String, String> = mutableMapOf(),
-        fileFieldName: String = "file",
-        fileName: String = uploadFile.name
-    ): Response {
-        val jwtManager = SpringContextUtil.getBean(JwtManager::class.java)
-        if (jwtManager.isSendEnable()) {
-            val jwtToken = jwtManager.getToken() ?: ""
-            headers[AUTH_HEADER_DEVOPS_JWT_TOKEN] = jwtToken
-        }
-        return OkhttpUtils.uploadFile(
-            url = url,
-            uploadFile = uploadFile,
-            headers = headers,
-            fileFieldName = fileFieldName,
-            fileName = fileName
-        )
-    }
-
-    fun doPostFromService(
-        url: String,
-        jsonParam: String,
-        headers: MutableMap<String, String> = mutableMapOf()
-    ): Response {
-        val jwtManager = SpringContextUtil.getBean(JwtManager::class.java)
-        if (jwtManager.isSendEnable()) {
-            val jwtToken = jwtManager.getToken() ?: ""
-            headers[AUTH_HEADER_DEVOPS_JWT_TOKEN] = jwtToken
-        }
-        return OkhttpUtils.doPost(url, jsonParam, headers)
-    }
-
-    fun downloadFileFromService(
-        url: String,
-        destPath: File,
-        headers: MutableMap<String, String> = mutableMapOf(),
-        connectTimeoutInSec: Long? = null,
-        readTimeoutInSec: Long? = null,
-        writeTimeoutInSec: Long? = null
-    ) {
-        val jwtManager = SpringContextUtil.getBean(JwtManager::class.java)
-        if (jwtManager.isSendEnable()) {
-            val jwtToken = jwtManager.getToken() ?: ""
-            headers[AUTH_HEADER_DEVOPS_JWT_TOKEN] = jwtToken
-        }
-        OkhttpUtils.downloadFile(
-            url = url,
-            destPath = destPath,
-            headers = headers,
-            connectTimeoutInSec = connectTimeoutInSec,
-            readTimeoutInSec = readTimeoutInSec,
-            writeTimeoutInSec = writeTimeoutInSec
-        )
     }
 
     /**
