+++ src/backend/ci/core/store/biz-store/src/main/kotlin/com/tencent/devops/store/common/utils/StoreFileAnalysisUtil.kt
@@ -40,8 +40,8 @@ import com.tencent.devops.common.api.util.JsonUtil
 import com.tencent.devops.common.api.util.UUIDUtil
 import com.tencent.devops.common.api.util.YamlUtil
 import com.tencent.devops.common.client.Client
-import com.tencent.devops.common.service.utils.CommonUtils
 import com.tencent.devops.common.service.utils.ZipUtil
+import com.tencent.devops.common.web.utils.CommonServiceUtils
 import com.tencent.devops.common.web.utils.I18nUtil
 import com.tencent.devops.store.constant.StoreMessageCode
 import com.tencent.devops.store.constant.StoreMessageCode.USER_UPLOAD_PACKAGE_INVALID
@@ -148,10 +148,10 @@ object StoreFileAnalysisUtil {
     ): Result<Boolean?> {
         val serviceUrlPrefix = client.getServiceUrl(ServiceArchiveComponentFileResource::class)
         val serviceUrl = "$serviceUrlPrefix/service/artifactories/store/component/pkg/archive" +
-                "?userId=$userId&storeType=${archiveStorePkgRequest.storeType.name}" +
-                "&storeCode=${archiveStorePkgRequest.storeCode}&version=${archiveStorePkgRequest.version}" +
-                "&releaseType=${archiveStorePkgRequest.releaseType.name}"
-        CommonUtils.uploadFileToService(
+            "?userId=$userId&storeType=${archiveStorePkgRequest.storeType.name}" +
+            "&storeCode=${archiveStorePkgRequest.storeCode}&version=${archiveStorePkgRequest.version}" +
+            "&releaseType=${archiveStorePkgRequest.releaseType.name}"
+        CommonServiceUtils.uploadFileToService(
             url = serviceUrl,
             uploadFile = file,
             headers = mutableMapOf(AUTH_HEADER_USER_ID to userId)
@@ -174,16 +174,18 @@ object StoreFileAnalysisUtil {
         file: File
     ): Result<Boolean?> {
         val serviceUrlPrefix = client.getServiceUrl(ServiceArchiveAtomFileResource::class)
-        val serviceUrl = StringBuilder("$serviceUrlPrefix/service/artifactories/archiveAtom?userId=$userId" +
+        val serviceUrl = StringBuilder(
+            "$serviceUrlPrefix/service/artifactories/archiveAtom?userId=$userId" +
                 "&projectCode=${archiveAtomRequest.projectCode}&atomCode=${archiveAtomRequest.atomCode}" +
-                "&version=${archiveAtomRequest.version}")
+                "&version=${archiveAtomRequest.version}"
+        )
         archiveAtomRequest.releaseType?.let {
             serviceUrl.append("&releaseType=${archiveAtomRequest.releaseType!!.name}")
         }
         archiveAtomRequest.os?.let {
             serviceUrl.append("&os=${archiveAtomRequest.os}")
         }
-        CommonUtils.uploadFileToService(serviceUrl.toString(), file).use { response ->
+        CommonServiceUtils.uploadFileToService(serviceUrl.toString(), file).use { response ->
             response.body!!.string()
             if (!response.isSuccessful) {
                 return I18nUtil.generateResponseDataObject(
